requirements.txt
antiorm==1.2.1
astroid==2.9.3
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
db==0.1.1
db-sqlite3==0.0.1
Flask==2.0.3
Flask-Cors==3.0.10
idna==3.3
isort==5.10.1
itsdangerous==2.1.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.6.1
platformdirs==2.5.1
pycodestyle==2.8.0
pylint==2.12.2
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
toml==0.10.2
typing-extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
wrapt==1.13.3
